+++ test/test_helper.rb
@@ -2,12 +2,12 @@
 
 require "simplecov"
 
-SimpleCov.external_at_exit = true
 SimpleCov.start do
   enable_coverage :branch
   track_files "lib/**/*.rb"
   add_filter "/test/"
-  minimum_coverage line: 95, branch: 95
+  add_filter "/version.rb"
+  minimum_coverage line: 99, branch: 99
 end
 
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)